tool/requirements.txt
numpy
openai
python-dotenv
pydantic
